+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 4.37.0 ##
+
+1.  `SaintsDictionary<,>` not support `interface` type & `abstruct` type directly 
+2.  Add `[KeyAttribute(type, arguments...)]` and `[ValueAttribute(type, arguments...)]` to inject attributes to key/value for `SaintsDictionary<,>`
+
 ## 4.36.0 ##
 
 1.  You can now directly serialize an interface field with extended serialization.